+++ src/infrastructure/prun-api/data/create-map-getter.ts
@@ -4,7 +4,7 @@ const upperCase = (value: string) => value.toUpperCase();
 
 export function createMapGetter<T>(
   items: Ref<T[] | undefined>,
-  selector: (item: T) => string | string[],
+  selector: (item: T) => string | null | undefined | (string | null | undefined)[],
   valueTransformer?: (value: string) => string,
 ) {
   valueTransformer ??= upperCase;
@@ -16,6 +16,9 @@ export function createMapGetter<T>(
     for (const item of items.value) {
       const values = castArray(selector(item));
       for (const value of values) {
+        if (value === null || value === undefined) {
+          continue;
+        }
         map.set(valueTransformer(value), item);
       }
     }
@@ -27,7 +30,7 @@ export function createMapGetter<T>(
 
 export function createGroupMapGetter<T>(
   items: Ref<T[] | undefined>,
-  selector: (item: T) => string,
+  selector: (item: T) => string | null | undefined,
   valueTransformer?: (value: string) => string,
 ) {
   valueTransformer ??= upperCase;
@@ -37,7 +40,11 @@ export function createGroupMapGetter<T>(
     }
     const map = new Map<string, T[]>();
     for (const item of items.value) {
-      const key = valueTransformer(selector(item));
+      const value = selector(item);
+      if (value === null || value === undefined) {
+        continue;
+      }
+      const key = valueTransformer(value);
       let group = map.get(key);
       if (!group) {
         group = [];